Editing for functionality and tests
