+++ src/cases/components/allocate-role/allocate-role.component.html
@@ -1,7 +1,3 @@
-
-  <a *ngIf="roleAllocation!=='Exclusion'" class="govuk-back-link">
-    Back
-  </a>
   <h1 class="govuk-heading-l" >
     <span class="govuk-caption-l">{{caption}}</span>
     {{title}}
@@ -24,7 +20,4 @@
       </div>
     </fieldset>
   </div>
-  <div class="govuk-form-group">
-    <button class="govuk-button govuk-!-margin-right-1" type="submit">Continue</button>
-      <p><a href="javascript:void(0)">Cancel</a></p>
-  </div>
+
