+++ src/cases/containers/allocate-role-container/allocate-role-container.component.spec.ts
@@ -1,5 +1,6 @@
-import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { CUSTOM_ELEMENTS_SCHEMA, DebugElement } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
 import { ActivatedRoute } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { Store } from '@ngrx/store';
@@ -11,6 +12,9 @@ import { RoleAllocationType } from '../../../cases/enums';
 import { RoleAllocationConstants } from '../../components/constants';
 
 import { AllocateRoleContainerComponent } from '..';
+import { DescribeExclusionComponent } from 'src/cases/components/describe-exclusion/describe-exclusion.component';
+import { ErrorMessageComponent } from 'src/app/components/error-message/error-message.component';
+import { ReactiveFormsModule } from '@angular/forms';
 
 describe('AllocateRoleContainerComponent', () => {
   let component: AllocateRoleContainerComponent;
@@ -23,6 +27,7 @@ describe('AllocateRoleContainerComponent', () => {
     TestBed.configureTestingModule({
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
       imports: [
+        ReactiveFormsModule,
         RouterTestingModule
       ],
       providers: [provideMockStore(),
@@ -35,7 +40,7 @@ describe('AllocateRoleContainerComponent', () => {
             },
           }
         }],
-      declarations: [AllocateRoleContainerComponent],
+      declarations: [AllocateRoleContainerComponent, DescribeExclusionComponent, ErrorMessageComponent],
     })
       .compileComponents();
   }));
@@ -46,6 +51,7 @@ describe('AllocateRoleContainerComponent', () => {
     fixture = TestBed.createComponent(AllocateRoleContainerComponent);
     component = fixture.componentInstance;
     spyOnPipeToStore.and.returnValue(of([{isCaseAllocator: true}, {}]));
+    component.describingExclusion = false;
     fixture.detectChanges();
 
   });
@@ -68,6 +74,40 @@ describe('AllocateRoleContainerComponent', () => {
       expect(component.roleAllocationType).toBe(RoleAllocationType.Exclusion);
     });
 
+  });
+
+  describe('describe exclusion', () => {
+    it('should submit form when all values are populated', () => {
+      spyOn(component, 'continue');
+      const container: DebugElement = fixture.debugElement.query(By.css('#submit-button'));
+      component.allocationForm.get(component.DESCRIBE_EXCLUSION_CONTROL_NAME).patchValue('some text');
+      const button: HTMLButtonElement = container.nativeElement as HTMLButtonElement;
+      button.click();
+      expect(component.continue).toHaveBeenCalledWith({text: 'some text'}, true);
+    });
+
+    it('should set the form to an invalid state if none of the values are populated and display error', () => {
+      spyOn(component, 'continue');
+      component.error = {
+        title: 'There is a problem',
+        description: 'Enter exclusion',
+        fieldId: 'exclusion-description'
+      };
+      fixture.detectChanges();
+      const buttonDebugElement: DebugElement = fixture.debugElement.query(By.css('#submit-button'));
+      const errorDebugElement: DebugElement = fixture.debugElement.query(By.css('.govuk-error-summary__list'));
+      const button: HTMLButtonElement = buttonDebugElement.nativeElement as HTMLButtonElement;
+      const error: HTMLElement = errorDebugElement.nativeElement as HTMLElement;
+      button.click();
+      expect(component.continue).toHaveBeenCalledWith({text: ''}, false);
+      expect(error.firstChild.firstChild.textContent).toBe('Enter exclusion');
+    });
+
+    it('should mark the form as touched when the user submits the form', () => {
+      spyOn(component.allocationForm.get('text'), 'markAsTouched');
+      component.continue({text: 'some text'}, true);
+      expect(component.allocationForm.get('text').markAsTouched).toHaveBeenCalled();
+    });
   });
 
 });
