--- src/cases/containers/role-container/allocate-role-container.component.html
@@ -1,15 +1,0 @@
-<form [formGroup]="allocationForm" (ngSubmit)="continue(allocationForm.value, allocationForm.valid)">
-  <a href="../" class="govuk-back-link">Back</a>
-  <main class="govuk-main-wrapper">
-    <exui-error-message *ngIf="error" [error]="error"></exui-error-message>
-    <exui-describe-exclusion
-      [submitted]="submitted"
-      [formGroup]="allocationForm"
-      [title]="'Add an exclusion'"
-      [description]="'Describe the exclusion'"
-      [controlName]="DESCRIBE_EXCLUSION_CONTROL_NAME"
-    ></exui-describe-exclusion>
-    <button id="submit-button" class="govuk-button govuk-!-margin-right-1" type="submit">Continue</button>
-    <p><a class="govuk-!-margin-top-3" id="cancel-link" href="javascript:void(0)">Cancel</a></p>
-  </main>
-</form>
